meta: move chaos rule and request types to rules.rs
